[Fix] AppContainerを廃止して依存関係はBuildConfigではなくアプリのtargetで注入するように
