+++ Package/Sources/Presentation/Screen/Home/HomeView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import Environment
 
 public struct HomeView: View {
