+++ Package/Sources/Presentation/Screen/Launch/LaunchView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import Environment
 
 public struct LaunchView: View {
