+++ Package/Sources/Presentation/Screen/Main/MainView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import Environment
 
 public struct MainView: View {
