+++ Package/Sources/Presentation/Screen/SearchResult/Component/TagInformationView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import Environment
 
 public struct TagInformationView: View {
