+++ Package/Sources/Presentation/Screen/Stock/StockView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import Environment
 
 public struct StockView: View {
